Refactor SettingItem into separate component
Also absorb the plain-JS parts into React
