+++ src/settings/components/SettingItem/SettingItem.module.css
@@ -1,0 +1,46 @@
+.settingItem {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    padding: 16px 0;
+    border-bottom: 1px solid #eee;
+}
+
+.settingItem:last-child {
+    border-bottom: none;
+}
+
+.settingLabel {
+    flex: 0 0 200px;
+    padding-right: 24px;
+}
+
+.settingLabel label {
+    margin: 0;
+    font-size: 14px;
+    font-weight: 500;
+    color: #333;
+}
+
+.settingLabel .description {
+    font-size: 12px;
+    color: #666;
+    margin-top: 4px;
+}
+
+.settingControl select {
+    width: 200px;
+    padding: 8px 12px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    font-size: 14px;
+    color: #333;
+    background-color: white;
+}
+
+.settingControl select:focus {
+    outline: none;
+    border-color: #0066cc;
+    box-shadow: 0 0 0 2px rgba(0,102,204,0.2);
+}
